% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha1.R
\name{sha1}
\alias{sha1}
\title{Calculate a SHA1 hash of an object.}
\usage{
sha1(x, digits = 14L, zapsmall = 7L)
}
\arguments{
\item{x}{The object to calculate the SHA1}

\item{digits}{The approximate number of significant digits in base 10. Will
be converted to a base 16 equivalent. Defaults to \code{digits = 14},
expect for sha1.anova where \code{digits = 4}}

\item{zapsmall}{The apporixmate negative magnitute of the smallest relevant
digit. Will be converted to a base 2 equivalent. Values smaller than this
number are equivalent to 0. Defaults to \code{zapsmall = 7}}
}
\description{
Calculate a SHA1 hash of an object. The main difference with
\code{digest(x, algo = "sha1")} is that \code{sha1()} will give the same hash
on 32-bit and 64-bit systems. Note that the results depends on the setting of
\code{digits} and \code{zapsmall} when handling floating point numbers. The
current defaults keep \code{digits} and \code{zapsmall} as large as possible
while maintaining the same hash on 32 bit and 64 bit systems.
}
\author{
Thierry Onkelinx
}

